Rewrite index compare in ReRoll to work
Old compare actually sent the meme if the index wasn't equal to the first index in oldIndexes. It now marks it as bad if matches any index in oldIndexes and rerolls until index does not match any. It also properly intializes oldIndexes without 0 and checks for 10  length after pushing the tenth index instead of before.