Restructure the code to use a stack of URLs.
Right at 4:20, the bot grabs the top 10 memes of the day and stores them
in an stack. Then, it randomizes the order of them to preserve the random
nature of the bot. Then, it posts the first URL from the array. On every
consecutive re-roll, it pops the next one off the stack until it's empty.
I also did some various clean up, like making smaller functions, renaming
things to be more clear, and deleting redundant code. This method of
getting new memes is easier to read and understand, and it implicitly
ensures that there are no repeats.